Refactoring and bug fixes

+++ backend/app/models/files.py
@@ -7,8 +7,8 @@ from app.extensions import db
 class File(db.Model, SerializerMixin):
     __tablename__ = 'file'
 
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BigInteger, primary_key=True)
     name = db.Column(db.String(150), nullable=False)
     telegram_file_id = db.Column(db.String(150), nullable=False)
-    user_id = db.Column(db.Integer, db.ForeignKey('base_user.id'), nullable=False)
+    user_id = db.Column(db.BigInteger, db.ForeignKey('base_user.id'), nullable=False)
