+++ backend/app/models/users.py
@@ -8,7 +8,7 @@ from app.extensions import db, bcrypt
 class BaseUser(db.Model, SerializerMixin):
     __tablename__ = 'base_user'
 
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BigInteger, primary_key=True)
     
     username = db.Column(db.String(50))
     password_hash = db.Column(db.String(255), nullable=False)    
@@ -28,11 +28,11 @@ class BaseUser(db.Model, SerializerMixin):
     
     
 class Role(db.Model, SerializerMixin):
-    id = db.Column(db.Integer, primary_key=True)
+    id = db.Column(db.BigInteger, primary_key=True)
     name = db.Column(db.String(80), unique=True)
     
 
 user_roles = db.Table('user_roles',
-    db.Column('user_id', db.Integer, db.ForeignKey('base_user.id'), primary_key=True),
-    db.Column('role_id', db.Integer, db.ForeignKey('role.id'), primary_key=True)
+    db.Column('user_id', db.BigInteger, db.ForeignKey('base_user.id'), primary_key=True),
+    db.Column('role_id', db.BigInteger, db.ForeignKey('role.id'), primary_key=True)
 )
